requirements.txt
streamlit
groq
forex-python
pint
python-dotenv
